*: move the code of distinct logic. (#2539)
