+++ plan/physical_plan_builder.go
@@ -32,7 +32,6 @@ const (
 	netWorkFactor   = 1.5
 	memoryFactor    = 5.0
 	selectionFactor = 0.8
-	distinctFactor  = 0.7
 	cpuFactor       = 0.9
 	aggFactor       = 0.1
 	joinFactor      = 0.3
